+++ src/PetHome.API/WeatherForecast.cs
@@ -1,13 +1,11 @@
-namespace PetHome.API
+namespace PetHome.API;
+public class WeatherForecast
 {
-    public class WeatherForecast
-    {
-        public DateOnly Date { get; set; }
+    public DateOnly Date { get; set; }
 
-        public int TemperatureC { get; set; }
+    public int TemperatureC { get; set; }
 
-        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
+    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
 
-        public string? Summary { get; set; }
-    }
+    public string? Summary { get; set; }
 }
