+++ src/PetHome.Domain/GeneralValueObjects/PaymentMethodEnum.cs
@@ -1,15 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PetHome.Domain.GeneralValueObjects
+﻿namespace PetHome.Domain.GeneralValueObjects;
+public enum PaymentMethodEnum
 {
-    public enum PaymentMethodEnum
-    {
-        Карта,
-        Наличные,
-        Помощь
-    }
+    Card,
+    Cash,
+    Help
 }
