+++ src/PetHome.Domain/GeneralValueObjects/PhoneNumber.cs
@@ -1,43 +1,30 @@
 ﻿using CSharpFunctionalExtensions;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Text.RegularExpressions;
-using System.Threading.Tasks;
 
-namespace PetHome.Domain.GeneralValueObjects
+namespace PetHome.Domain.GeneralValueObjects;
+public class PhoneNumber : ValueObject
 {
+    private const string PhoneNumberRegex = @"^\(?([0-9]{3})\)?[-.●]?([0-9]{3})[-.●]?([0-9]{4})$";
+    private string Number { get; }
 
-    public class PhoneNumber : ValueObject
+    private PhoneNumber(string number)
     {
-        private const string PhoneNumberRegex = @"^\(?([0-9]{3})\)?[-.●]?([0-9]{3})[-.●]?([0-9]{4})$";
-
-        private string Number { get; }
-
-        private PhoneNumber(string number)
-        {
-            Number = number;
-        }
-
-        public Result<PhoneNumber> Create(string number)
-        {
-            if (string.IsNullOrWhiteSpace(number))
-                return Result.Failure<PhoneNumber>("Введите номер телефона");
-
-            if (!Regex.IsMatch(number, PhoneNumberRegex))
-                return Result.Failure<PhoneNumber>("Номер телефона не соответствует формату");
-
-
-            return new PhoneNumber(number);
-        }
-
-
-
-        protected override IEnumerable<object> GetEqualityComponents()
-        {
-            yield return Number;
-        }
+        Number = number;
     }
 
+    public Result<PhoneNumber> Create(string number)
+    {
+        if (string.IsNullOrWhiteSpace(number))
+            return Result.Failure<PhoneNumber>("Введите номер телефона");
+
+        if (!Regex.IsMatch(number, PhoneNumberRegex))
+            return Result.Failure<PhoneNumber>("Номер телефона не соответствует формату");
+
+        return new PhoneNumber(number);
+    }
+
+    protected override IEnumerable<object> GetEqualityComponents()
+    {
+        yield return Number;
+    }
 }
