+++ src/PetHome.Domain/GeneralValueObjects/Requisites.cs
@@ -1,50 +1,32 @@
 ﻿using CSharpFunctionalExtensions;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace PetHome.Domain.GeneralValueObjects
+namespace PetHome.Domain.GeneralValueObjects;
+public class Requisites : ValueObject
 {
+    private string Name { get; }
+    private string Description { get; }
+    private PaymentMethodEnum PaymentMethod { get; }
 
-
-    public class Requisites : ValueObject
+    private Requisites(string name, string description, PaymentMethodEnum paymentMethod)
     {
-        private string Name { get; }
-        private string Description { get; }
-        private PaymentMethodEnum PaymentMethod { get; }
+        Name = name;
+        Description = description;
+        PaymentMethod = paymentMethod;
+    }
 
+    public static Result<Requisites> Create(string name, string description, PaymentMethodEnum paymentMethod)
+    {
+        bool isInvalidRequisite = string.IsNullOrEmpty(name) || string.IsNullOrEmpty(description) || paymentMethod == null;
 
+        if (isInvalidRequisite)
+            return Result.Failure<Requisites>("Некорректно введены реквезиты");
 
-        private Requisites(string name, string description, PaymentMethodEnum paymentMethod)
-        {
-            Name = name;
-            Description = description;
-            PaymentMethod = paymentMethod;
-        }
+        return new Requisites(name, description, paymentMethod);
+    }
 
-
-        public static Result<Requisites> Create(string name, string description, PaymentMethodEnum paymentMethod)
-        {
-            bool isInvalidRequisite = string.IsNullOrEmpty(name) || string.IsNullOrEmpty(description) || paymentMethod == null;
-
-
-            if (isInvalidRequisite)
-                return Result.Failure<Requisites>("Некорректно введены реквезиты");
-
-
-            return new Requisites(name, description, paymentMethod);
-        }
-
-
-
-        protected override IEnumerable<object> GetEqualityComponents()
-        {
-            yield return Name;
-            yield return PaymentMethod;
-        }
-
-
+    protected override IEnumerable<object> GetEqualityComponents()
+    {
+        yield return Name;
+        yield return PaymentMethod;
     }
 }
